requirements.txt
streamlit==1.32.0
pandas==2.2.2
numpy==1.26.4
plotly==5.21.0
scikit-learn==1.4.2
tensorflow-cpu==2.15.0
vnstock
joblib